cpp - renamied ordered circular queue
